feat: Rename all files to kebab-case
- Rename all component files from PascalCase to kebab-case
- Update all import statements to use new kebab-case file names
- Update index.ts export statements
- Fix relative imports within components
- All builds passing ✓

This resolves git case-sensitivity issues and improves consistency.
