+++ src/app/articles/page.tsx
@@ -1,4 +1,4 @@
-import { ArticlesBody } from '@/components/pages/articles/ArticlesBody'
+import { ArticlesBody } from '@/components/pages/articles/articles-body'
 
 import { setMetaData } from '@/utils'
 import { fetchArticles } from '@/libs'
