+++ src/app/contact/thanks/page.tsx
@@ -1,5 +1,5 @@
-import { Title } from '@/components/common/Title'
-import { TextBtn } from '@/components/common/TextBtn'
+import { Title } from '@/components/common/title'
+import { TextBtn } from '@/components/common/text-btn'
 
 import { setMetaData } from '@/utils'
