+++ src/components/common/category-icon.tsx
