+++ src/components/common/image-wrapper.tsx
