+++ src/components/pages/articles/articles-list.tsx
@@ -1,5 +1,5 @@
-import { CategoryIcon } from '@/components/common/CategoryIcon'
-import { ImageWrapper } from '@/components/common/ImageWrapper'
+import { CategoryIcon } from '@/components/common/category-icon'
+import { ImageWrapper } from '@/components/common/image-wrapper'
 
 import { FeedObj } from '@/types'
