+++ src/components/pages/works/works-detail-info.tsx
