+++ src/libs/fetch/fetch-works-detail.ts
