+++ src/utils/index.ts
@@ -1,7 +1,7 @@
-export { conversionDate } from './conversionDate'
-export { cutText } from './cutText'
-export { isEmptyObject } from './isEmptyObject'
-export { loadScript } from './loadScript'
-export { removeHtml } from './removeHtml'
-export { setMetaData } from './setMetadata'
+export { conversionDate } from './conversion-date'
+export { cutText } from './cut-text'
+export { isEmptyObject } from './is-empty-object'
+export { loadScript } from './load-script'
+export { removeHtml } from './remove-html'
+export { setMetaData } from './set-metadata'
 export { shuffle } from './shuffle'
