+++ src/utils/is-empty-object.ts
