+++ src/utils/load-script.ts
